move fix_module_version to environment.yml file and fix test fixture
